fix(music): resolve PR review issues in scrobbling integration
- Fix TokenStore API: replace .set() with .save() for StringSetting
- Fix LastFMTokenStore.store() signature to accept (sessionKey, username)
- Add retrieveSessionKey() and retrieveUsername() methods to LastFMTokenStore
- Move scrobbling UI registration (LB + Last.fm categories, Scrobbling sub-screen)
  from SettingsPatch.kt into ListenBrainzPatch.kt (scrobblingPatch) so it only
  activates when the patch is applied
- Fix NonInteractivePreference key collision: rename LB and Last.fm token UI
  preference keys to *_token_ui to avoid clashing with registered StringSettings
  and prevent 'unhandled setting' popup in AbstractPreferenceFragment
