+++ patches/src/main/kotlin/app/morphe/patches/music/misc/settings/SettingsPatch.kt
@@ -147,83 +147,6 @@ val settingsPatch = bytecodePatch(
             SwitchPreference("morphe_show_menu_icons")
         )
 
-        val listenBrainzCategory = PreferenceCategory(
-            key = "morphe_settings_music_screen_5_listenbrainz",
-            preferences = setOf(
-                NonInteractivePreference(
-                    key = "morphe_music_listenbrainz_token",
-                    titleKey = "morphe_music_listenbrainz_token_title",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.music.settings.preference.ListenBrainzTokenPreference",
-                    selectable = true
-                ),
-                SwitchPreference("morphe_music_listenbrainz_enabled"),
-                SwitchPreference("morphe_music_listenbrainz_now_playing"),
-                NonInteractivePreference(
-                    key = "morphe_music_listenbrainz_min_song_duration",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.shared.settings.preference.SeekBarPreference",
-                    selectable = true
-                ),
-                NonInteractivePreference(
-                    key = "morphe_music_listenbrainz_delay_percent",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.shared.settings.preference.SeekBarPreference",
-                    selectable = true
-                ),
-                NonInteractivePreference(
-                    key = "morphe_music_listenbrainz_delay_seconds",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.shared.settings.preference.SeekBarPreference",
-                    selectable = true
-                )
-            )
-        )
-
-        val lastfmCategory = PreferenceCategory(
-            key = "morphe_music_lastfm_category",
-            titleKey = "morphe_settings_music_lastfm_title",
-            preferences = setOf(
-                NonInteractivePreference(
-                    key = "morphe_music_lastfm_token",
-                    titleKey = "morphe_music_lastfm_token_title",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.music.settings.preference.LastFMTokenPreference",
-                    selectable = true
-                ),
-                SwitchPreference("morphe_music_lastfm_enabled"),
-                SwitchPreference("morphe_music_lastfm_now_playing"),
-                NonInteractivePreference(
-                    key = "morphe_music_lastfm_min_song_duration",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.shared.settings.preference.SeekBarPreference",
-                    selectable = true
-                ),
-                NonInteractivePreference(
-                    key = "morphe_music_lastfm_delay_percent",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.shared.settings.preference.SeekBarPreference",
-                    selectable = true
-                ),
-                NonInteractivePreference(
-                    key = "morphe_music_lastfm_delay_seconds",
-                    summaryKey = null,
-                    tag = "app.morphe.extension.shared.settings.preference.SeekBarPreference",
-                    selectable = true
-                )
-            )
-        )
-
-        val scrobblingScreen = PreferenceScreenPreference(
-            key = "morphe_settings_music_screen_5_scrobbling",
-            titleKey = "morphe_settings_music_screen_scrobbling_title",
-            summaryKey = "morphe_settings_music_screen_scrobbling_summary",
-            preferences = setOf(
-                listenBrainzCategory,
-                lastfmCategory
-            )
-        )
-
         PreferenceScreen.MISC.addPreferences(
             TextPreference(
                 key = null,
@@ -231,8 +154,7 @@ val settingsPatch = bytecodePatch(
                 summaryKey = "morphe_pref_import_export_summary",
                 inputType = InputType.TEXT_MULTI_LINE,
                 tag = "app.morphe.extension.shared.settings.preference.ImportExportPreference",
-            ),
-            scrobblingScreen
+            )
         )
 
         modifyActivityForSettingsInjection(
